# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Survivor Guessing Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
  /users/me/role:
    patch:
      operationId: updateMyRole
      tags: [users]
      summary: Update current user role (used during registration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameBody"
      responses:
        "201":
          description: Created game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{gameId}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a game by ID
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game (admin only)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameBody"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
  /games/{gameId}/weeks:
    get:
      operationId: listWeeks
      tags: [weeks]
      summary: List weeks for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of weeks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Week"
    post:
      operationId: createWeek
      tags: [weeks]
      summary: Create a week (admin only)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeekBody"
      responses:
        "201":
          description: Created week
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
  /weeks/{weekId}:
    get:
      operationId: getWeek
      tags: [weeks]
      summary: Get a week by ID
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Week
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Week"
  /weeks/{weekId}/questions:
    get:
      operationId: listQuestions
      tags: [questions]
      summary: List questions for a week
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions with choices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionWithChoices"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Create a question (admin only)
      parameters:
        - name: weekId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Created question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionWithChoices"
  /questions/{questionId}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionWithChoices"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /questions/{questionId}/choices:
    post:
      operationId: createChoice
      tags: [choices]
      summary: Add a choice to a question (admin only)
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChoiceBody"
      responses:
        "201":
          description: Created choice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Choice"
  /choices/{choiceId}:
    delete:
      operationId: deleteChoice
      tags: [choices]
      summary: Delete a choice (admin only)
      parameters:
        - name: choiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /games/{gameId}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get leaderboard for a game
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /games/{gameId}/stats:
    get:
      operationId: getGameStats
      tags: [games]
      summary: Get game stats summary
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Game stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStats"
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, player]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - username
        - role
        - createdAt
    UpdateRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [admin, player]
      required:
        - role
    Game:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [setup, active, completed]
        survivorWinnerContestantId:
          type: ["integer", "null"]
        currentWeekNumber:
          type: integer
        totalWeeks:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - status
        - survivorWinnerContestantId
        - currentWeekNumber
        - totalWeeks
        - createdAt
    CreateGameBody:
      type: object
      properties:
        name:
          type: string
        totalWeeks:
          type: integer
      required:
        - name
        - totalWeeks
    UpdateGameBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [setup, active, completed]
...
    Week:
      type: object
      properties:
        id:
          type: integer
        gameId:
          type: integer
        weekNumber:
          type: integer
        isLocked:
          type: boolean
        isOpen:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - gameId
        - weekNumber
        - isLocked
        - isOpen
        - createdAt
    CreateWeekBody:
      type: object
      properties:
        weekNumber:
          type: integer
      required:
        - weekNumber
    Choice:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        choiceText:
          type: string
      required:
        - id
        - questionId
        - choiceText
    QuestionWithChoices:
      type: object
      properties:
        id:
          type: integer
        weekId:
          type: integer
        text:
          type: string
        pointValue:
          type: integer
        choices:
          type: array
          items:
            $ref: "#/components/schemas/Choice"
      required:
        - id
        - weekId
        - text
        - pointValue
        - choices
    CreateQuestionBody:
      type: object
      properties:
        text:
          type: string
        pointValue:
          type: integer
        choices:
          type: array
          items:
            type: string
      required:
        - text
        - pointValue
    UpdateQuestionBody:
      type: object
      properties:
        text:
          type: string
        pointValue:
          type: integer
    CreateChoiceBody:
      type: object
      properties:
        choiceText:
          type: string
      required:
        - choiceText
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        totalPoints:
          type: integer
        weeklyPoints:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyPoints"
        survivorPickPoints:
          type: integer
      required:
        - rank
        - userId
        - username
        - totalPoints
        - weeklyPoints
        - survivorPickPoints
    WeeklyPoints:
      type: object
      properties:
        weekNumber:
          type: integer
        points:
          type: integer
      required:
        - weekNumber
        - points
    GameStats:
      type: object
      properties:
        totalPlayers:
          type: integer
        lockedWeeks:
          type: integer
        openWeeks:
          type: integer
        topPlayer:
          type: ["string", "null"]
        topPlayerPoints:
          type: ["integer", "null"]
      required:
        - totalPlayers
        - lockedWeeks
        - openWeeks
        - topPlayer
        - topPlayerPoints