# lib/api-zod/src/generated/types/questionWithChoices.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Survivor Guessing Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Choice } from "./choice";

export interface QuestionWithChoices {
  id: number;
  weekId: number;
  text: string;
  pointValue: number;
  choices: Choice[];
}