# lib/api-zod/src/generated/types/weeklyPoints.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Survivor Guessing Game API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyPoints {
  weekNumber: number;
  points: number;
}